# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tool kit test",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Swagger API server for Tool kit test"
        }
    ],
    "paths": {
        "/api": {},
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth user",
                "description": "Auth users",
                "operationId": "7f7ee36c4776c58183ce95c651310e11",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@mail.ru",
                                    "password": "1235"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "e7eb65de3f9c881e32bded41ee971e77",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh token",
                "description": "Refresh token",
                "operationId": "59a5c113caf93287a81a95ffba8d982c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/auth/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration user",
                "description": "Registration users",
                "operationId": "d9bedfca4284c126e71faec66247c00b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@mail.ru",
                                    "password": "1235",
                                    "address": "Moscow",
                                    "phone": "+79999999999",
                                    "birthday": "2023-06-03 23:32:17"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/statement": {
            "get": {
                "tags": [
                    "Statment"
                ],
                "summary": "Get statements list",
                "description": "Display a listing of the resource.",
                "operationId": "cb1090f5b6212c8f9bd9ce76e55fe12f",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Statment"
                ],
                "summary": "Add statement",
                "description": "Store a newly created resource in storage.",
                "operationId": "527f12a5957b7bc8c99d0db0d1f54bce",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "file": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Name",
                                    "number": "1235",
                                    "file": "C://file.txt"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/statement/{id}": {
            "get": {
                "tags": [
                    "Statment"
                ],
                "summary": "Get statement by id",
                "description": "Display the specified resource.",
                "operationId": "25896756304319ec0c001f9349fa91c1",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Statment"
                ],
                "summary": "Update statment by id",
                "description": "Update the specified resource in storage.",
                "operationId": "6c8f64a9c3ceb97607722623079e2c27",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Name",
                                    "number": "1235",
                                    "file": "C://file.txt"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Statment"
                ],
                "summary": "Delete statment by id",
                "description": "Remove the specified resource from storage.",
                "operationId": "25ad435d8d4083a3c408bf4fe42fc302",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 15
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}